ENGR00213549 Add MenuInfo check before create context menu
The mouse right button click on Music UI will trigger a
context menu create on List. But the activities did not
handle the null ContextMenuInfo param which is triggered
by right click on the empty list. This causes crash.
The music app handle the finger event well, but not mouse.

Signed-off-by: Xinyu Chen <xinyu.chen@freescale.com>